+++ internal/cli/tcp_cmd.go
@@ -49,7 +49,7 @@ func runTCP(ctx context.Context, args []string, configPath string, stdout, stder
 		fs.PrintDefaults()
 	}
 
-	if err := fs.Parse(args); err != nil {
+	if err := parseInterspersedFlags(fs, args); err != nil {
 		return 2
 	}
 	if fs.NArg() != 1 {